Faster `gem_version`
Implement `bundler.gem_version` to be constant time lookup using a hash instead of scanning with a detect.

Checking for Gemfile.lock is redundant since we have a `NoLockfile` language pack that detects this before any other packs can be run.

Reduce the Language Pack footprint so it knows less about the internals of the lockfile parser